Tweak when we send language to backend and retry if it fails
